+++ src/operators/index.ts
@@ -14,7 +14,7 @@ export * from './OBJECT_PROPERTIES'
 export * from './STRING_SUBSTITUTION'
 export * from './COUNT'
 export * from './SPLIT'
-export * from './PG_SQL'
+export * from './SQL'
 export * from './GRAPHQL'
 export * from './GET'
 export * from './POST'
