+++ src/operators/operatorAliases.ts
@@ -65,11 +65,12 @@ export const operatorAliases: OperatorAliases = {
   get: 'GET',
   api: 'GET',
   post: 'POST',
-  pgSql: 'PG_SQL',
-  sql: 'PG_SQL',
-  postgres: 'PG_SQL',
-  pg: 'PG_SQL',
-  pgDb: 'PG_SQL',
+  sql: 'SQL',
+  pgSql: 'SQL',
+  postgres: 'SQL',
+  pg: 'SQL',
+  sqLite: 'SQL',
+  sqlite: 'SQL',
   graphQl: 'GRAPHQL',
   graphql: 'GRAPHQL',
   gql: 'GRAPHQL',
